+++ resources/views/admin/sizes/index.blade.php
